+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2020-02-18
+
+### Added
+
+- The `update_taxon_assignments` method was added to allow users to create,
+  replace, or remove taxon assignments for a Genome object.
+
 ## [0.10.4] - 2020-02-20
 
 ### Changed